build refactor & code movement
