+++ cutlass/nvf_cutlass.h
@@ -11,6 +11,31 @@
 
 namespace nvfuser::cutlass_kernels {
 
+// Validates all input parameters and tensor properties for NVFP4 scaled matrix
+// multiplication
+//
+// This function performs comprehensive validation of input tensors including:
+// - CUDA device and contiguity checks
+// - Data type validation for all inputs
+// - Matrix dimension and shape compatibility
+// - Alignment requirements for optimal performance
+// - Scale matrix shape validation
+//
+// Parameters:
+//   a, b: Input matrices to validate
+//   scales_a, scales_b: Scale matrices to validate
+//   alpha: Alpha scaling factor to validate
+//
+// Returns: Tuple of (m, n, k) dimensions for the GEMM operation
+//
+// Throws: NVF_CHECK exceptions for any validation failures
+std::tuple<int64_t, int64_t, int64_t> validateInputsNvfp4ScaledMm(
+    const torch::Tensor& a,
+    const torch::Tensor& b,
+    const torch::Tensor& scales_a,
+    const torch::Tensor& scales_b,
+    const torch::Tensor& alpha);
+
 // Performs scaled matrix multiplication using NVFP4 format
 //
 // This function implements a scaled matrix multiplication C = alpha * (A @ B)
