+++ cutlass/nvfp4_scaled_mm.cu
@@ -276,25 +276,9 @@ inline int64_t roundUp(int64_t x, int64_t y) {
   return (x + y - 1) / y * y;
 }
 
-// Validates all input parameters and tensor properties for NVFP4 scaled matrix
-// multiplication
-//
-// This function performs comprehensive validation of input tensors including:
-// - CUDA device and contiguity checks
-// - Data type validation for all inputs
-// - Matrix dimension and shape compatibility
-// - Alignment requirements for optimal performance
-// - Scale matrix shape validation
-//
-// Parameters:
-//   a, b: Input matrices to validate
-//   scales_a, scales_b: Scale matrices to validate
-//   alpha: Alpha scaling factor to validate
-//
-// Returns: Tuple of (m, n, k) dimensions for the GEMM operation
-//
-// Throws: NVF_CHECK exceptions for any validation failures
-std::tuple<int64_t, int64_t, int64_t> validateInputs(
+} // namespace
+
+std::tuple<int64_t, int64_t, int64_t> validateInputsNvfp4ScaledMm(
     const torch::Tensor& a,
     const torch::Tensor& b,
     const torch::Tensor& scales_a,
@@ -396,8 +380,6 @@ std::tuple<int64_t, int64_t, int64_t> validateInputs(
   return {m, n, k};
 }
 
-} // namespace
-
 torch::Tensor nvfp4_scaled_mm(
     const torch::Tensor& a,
     const torch::Tensor& b,
@@ -406,7 +388,7 @@ torch::Tensor nvfp4_scaled_mm(
     const torch::Tensor& alpha,
     const at::ScalarType out_dtype) {
   // Validate all inputs and get matrix dimensions
-  auto [m, n, k] = validateInputs(a, b, scales_a, scales_b, alpha);
+  auto [m, n, k] = validateInputsNvfp4ScaledMm(a, b, scales_a, scales_b, alpha);
 
   at::cuda::CUDAGuard device_guard{(int8_t)a.get_device()};
   const cudaStream_t stream = at::cuda::getCurrentCUDAStream(a.get_device());
