+++ tests/python/utils/__init__.py
@@ -2,5 +2,6 @@
 # All rights reserved.
 # SPDX-License-Identifier: BSD-3-Clause
 
-from . import utils  # noqa: F401,F403
+from . import utils, narrow_precision  # noqa: F401,F403
 from .utils import *  # noqa: F401,F403
+from .narrow_precision import *  # noqa: F401,F403
