+++ tests/python_direct/test_cutlass_nvfp4_gemm.py
@@ -13,78 +13,12 @@ if torch.cuda.get_device_capability() < (10, 0):
         allow_module_level=True,
     )
 
-FLOAT4_E2M1_MAX = 6.0
-FLOAT8_E4M3_EPS = torch.finfo(torch.float8_e4m3fn).tiny
-FLOAT8_E4M3_MAX = torch.finfo(torch.float8_e4m3fn).max
-
-# Map the 7 values of e2m1 to corresponding positive fp32 value
-kE2M1ToFloatArray = [
-    0.0,
-    0.5,
-    1.0,
-    1.5,
-    2.0,
-    3.0,
-    4.0,
-    6.0,
-]
-
-
-# Convert FP4 into FP32
-def e2m1_to_fp32(int4_value):
-    signBit = int4_value & 0x8
-    int4_absValue = int4_value & 0x7
-    float_result = kE2M1ToFloatArray[int4_absValue]
-    if signBit:
-        float_result = -float_result
-    return float_result
-
-
-# Unpack float4_e2m1fn_x2 into two separate fp32 values
-def unpack_fp4_bytes(a, dtype):
-    assert a.dtype == torch.float4_e2m1fn_x2
-    m, n = a.shape
-    a = a.view(torch.uint8).flatten()
-    upper_half_byte = (a & 0xF0) >> 4
-    lower_half_byte = a & 0x0F
-    upper_half_float = torch.tensor([e2m1_to_fp32(x) for x in upper_half_byte]).to(
-        a.device
-    )
-    lower_half_float = torch.tensor([e2m1_to_fp32(x) for x in lower_half_byte]).to(
-        a.device
-    )
-    out = torch.stack((lower_half_float, upper_half_float), dim=-1).reshape(m, n * 2)
-    return out
-
-
-# Convert from packed block layout into general block-scale factor
-def convert_swizzled_to_linear(a_sf_swizzled: torch.Tensor, m, k, block_size):
-    sf_m, sf_k = a_sf_swizzled.shape
-    m_tiles = (m + 128 - 1) // 128
-    f = block_size * 4
-    k_tiles = (k + f - 1) // f
-    tmp = torch.reshape(a_sf_swizzled, (1, m_tiles, k_tiles, 32, 4, 4))
-    tmp = torch.permute(tmp, (0, 1, 4, 3, 2, 5))
-    return tmp.reshape(m_tiles * 128, k_tiles * f // block_size)
-
-
-def dequantize_to_dtype(
-    tensor_fp4, tensor_sf, global_scale, dtype, device, block_size=16
-):
-    """Dequantize the fp4 tensor back to high precision."""
-    # Two fp4 values are packed into one uint8.
-    assert tensor_fp4.dtype == torch.float4_e2m1fn_x2
-    m, packed_k = tensor_fp4.shape
-    k = packed_k * 2
-    tensor_f32 = unpack_fp4_bytes(tensor_fp4, dtype)
-    tensor_f32 = tensor_f32.reshape(m, k // block_size, block_size)
-    tensor_sf = tensor_sf.view(torch.float8_e4m3fn)
-    tensor_sf = convert_swizzled_to_linear(tensor_sf, m, k, block_size)
-    tensor_sf_dtype = tensor_sf.to(torch.float32) / global_scale
-
-    # scale the tensor
-    out = (tensor_f32 * tensor_sf_dtype.unsqueeze(-1)).reshape(m, k)
-    return out
+from python.utils import (
+    FLOAT4_E2M1_MAX,
+    FLOAT8_E4M3_MAX,
+    dequantize_to_dtype,
+    pytorch_nvfp4_quantize,
+)
 
 
 def get_ref_results(
@@ -110,53 +44,6 @@ def get_ref_results(
         b_fp4, b_sf, b_global_scale, dtype=dtype, device=device, block_size=block_size
     )
     return torch.matmul(a_in_dtype, b_in_dtype.t())
-
-
-# NOTE: This is from pytorch nvfp4 gemm tests.
-def to_fp4(x):
-    def down_size(size):
-        assert size[-1] % 2 == 0, f"{size} last dim not divisible by two"
-        return (*size[:-1], size[-1] // 2)
-
-    def pack_uint4(uint8_data) -> torch.Tensor:
-        # converting to uint8 for operations
-        shape = uint8_data.shape
-        assert shape[-1] % 2 == 0
-        uint8_data = uint8_data.contiguous().view(-1)
-        return (uint8_data[1::2] << 4 | uint8_data[::2]).view(down_size(shape))
-
-    from torch.testing._internal.common_quantized import _f32_to_floatx_unpacked
-
-    x = _f32_to_floatx_unpacked(x.float(), ebits=2, mbits=1)
-    x = pack_uint4(x)
-    x = x.view(torch.float4_e2m1fn_x2)
-    return x
-
-
-def pytorch_nvfp4_quantize(a, a_global_scale):
-    BLOCK_SIZE = 16
-    assert (
-        a.size(-1) % BLOCK_SIZE == 0
-    ), "The inner-most dim must be divisible by block_size; Padding is not implemented."
-    assert a.is_contiguous(), "Only contiguous tensors are supported."
-
-    original_shape = a.shape
-    a_fp32 = a.float().reshape(original_shape[0], -1, BLOCK_SIZE)
-
-    # Find absolute maximum along blockwise dimension
-    max_abs = torch.amax(torch.abs(a_fp32), dim=-1)
-    block_scale_fp32 = (max_abs / FLOAT4_E2M1_MAX).float()
-
-    scaled_block_scale_fp32 = block_scale_fp32 * a_global_scale
-    scaled_block_scale_fp8 = torch.clamp(
-        scaled_block_scale_fp32, min=FLOAT8_E4M3_EPS, max=FLOAT8_E4M3_MAX
-    ).to(torch.float8_e4m3fn)
-    scaled_block_scale_fp8_fp32 = scaled_block_scale_fp8.to(torch.float)
-    total_scale = scaled_block_scale_fp8_fp32 / a_global_scale
-    a_scaled = a_fp32 / total_scale.unsqueeze(-1)
-    a_scaled = torch.clamp(a_scaled, -FLOAT4_E2M1_MAX, FLOAT4_E2M1_MAX)
-    a_scaled = a_scaled.view(original_shape)
-    return to_fp4(a_scaled), scaled_block_scale_fp8
 
 
 @pytest.mark.parametrize("dtype", [torch.float16, torch.bfloat16])
